requirements.txt
# cat requirements.txt
robotframework==5.0.1
requests==2.27.1
unidecode==1.3.4
selenium==3.141.0
robotframework-seleniumlibrary==5.1.3
robotframework-browser==13.0.0
robotframework-datadriver==1.6.1
robotframework-pabot==2.5.3
robotframework-requests==0.9.3
robotframework-jsonlibrary==0.3.1
robotframework-imaplibrary2==0.4.0
robotframework-robocop==2.0.2
rpaframework==15.0.0
